+++ morpheus-graphql-core/src/Data/Morpheus/Types/App.hs
@@ -53,9 +53,9 @@ import Data.Morpheus.Types.Internal.Config
   )
 import Data.Morpheus.Types.Internal.Resolving
   ( ResolverContext (..),
+    ResolverRootValue,
     ResponseStream,
     ResultT (..),
-    RootResModel,
     cleanEvents,
     resultOr,
     runRootResModel,
@@ -64,7 +64,7 @@ import Data.Morpheus.Types.Internal.Stitching (Stitching (..))
 import Data.Morpheus.Validation.Document.Validation (ValidateSchema (..))
 import Relude hiding (empty)
 
-mkApp :: ValidateSchema s => Schema s -> RootResModel e m -> App e m
+mkApp :: ValidateSchema s => Schema s -> ResolverRootValue e m -> App e m
 mkApp appSchema appResolvers =
   resultOr
     FailApp
@@ -87,7 +87,7 @@ instance Monad m => Semigroup (App e m) where
 
 data AppData event (m :: * -> *) s = AppData
   { appConfig :: Config,
-    appResolvers :: RootResModel event m,
+    appResolvers :: ResolverRootValue event m,
     appSchema :: Schema s
   }