+++ morpheus-graphql-core/changelog.md
@@ -8,6 +8,7 @@
 
 ### Breaking changes
 
+- `RootResModel` was renamed to `ResolverRootValue`
 - `ResModel` was replaced with more general `ResolverValue`
 - `GQLScalar` was replaced with `EncodeScalar` and `DecodeScalar` type-classes.